pyproject.toml
[project]
name = "prodution-ready-rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google==3.0.0",
    "google-genai>=1.60.0",
    "groq>=1.0.0",
    "inngest>=0.5.14",
    "llama-index-core>=0.14.13",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "streamlit>=1.53.0",
    "uvicorn>=0.40.0",
]
